requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dataclasses==0.8
discord-ext-ipc==2.0.0
discord.py==1.7.3
filelock==3.0.12
greenlet==1.1.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.5.0
joblib==1.0.1
multidict==5.1.0
nltk==3.6.2
numpy==1.19.5
packaging==20.9
Pillow==8.2.0
pkg-resources==0.0.0
pyparsing==2.4.7
python-dotenv==0.17.1
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
sentencepiece==0.1.96
six==1.16.0
SQLAlchemy==1.4.19
tokenizers==0.8.0rc4
torch==1.9.0+cpu
torchaudio==0.9.0
torchvision==0.10.0+cpu
tqdm==4.61.1
transformers==3.0.0
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
zipp==3.4.1
